+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+
+## [2.1.0] - 2020-02-04
 ### Added
 - Bundle Lerna with the manager
 
@@ -156,7 +160,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]:    https://github.com/absolunet/node-manager/compare/2.0.1...HEAD
+[Unreleased]:    https://github.com/absolunet/node-manager/compare/2.1.0...HEAD
+[2.1.0]:         https://github.com/absolunet/node-manager/compare/2.0.1...2.1.0
 [2.0.1]:         https://github.com/absolunet/node-manager/compare/2.0.0...2.0.1
 [2.0.0]:         https://github.com/absolunet/node-manager/compare/2.0.0-rc.1...2.0.0
 [2.0.0-rc.1]:    https://github.com/absolunet/node-manager/compare/2.0.0-beta.1...2.0.0-rc.1